refactor: modularize firebase functions
