+++ functions/package.json
@@ -1,7 +1,7 @@
 {
   "name": "functions",
   "description": "Cloud Functions for Firebase",
-  "main": "lib/index.js",
+  "main": "lib/src/index.js",
   "scripts": {
     "clean": "rimraf lib",
     "build": "npm run clean && tsc",
