+++ functions/src/index.ts
@@ -1,0 +1,42 @@
+import * as admin from 'firebase-admin';
+
+if (!admin.apps.length) {
+  admin.initializeApp();
+}
+
+export { incrementReligionPoints, awardPointsToUser } from './orgHandlers';
+export {
+  completeChallenge,
+  createMultiDayChallenge,
+  completeChallengeDay,
+  generateChallenge,
+  generateDailyChallenge,
+  skipDailyChallenge,
+} from './challengeHandlers';
+export { askGeminiSimple, confessionalAI, askGeminiV2 } from './journalHandlers';
+export {
+  updateStreakAndXP,
+  getUserProfile,
+  seedFirestore,
+  backfillUserProfiles,
+  updateUserProfileCallable,
+  completeSignupAndProfile,
+} from './userHandlers';
+export {
+  startSubscriptionCheckout,
+  startOneTimeTokenCheckout,
+  startTokenCheckout,
+  createCheckoutSession,
+  createStripeSubscriptionIntent,
+  startDonationCheckout,
+  startCheckoutSession,
+  createStripeCheckout,
+  createStripeSetupIntent,
+  finalizePaymentIntent,
+  createTokenPurchaseSheet,
+  createSubscriptionSession,
+} from './stripeHandlers';
+
+export { onCompletedChallengeCreate } from '../firestoreArchitecture';
+export { handleStripeWebhookV2 } from '../stripeWebhooks';
+export { cleanLegacySubscriptionFields } from '../cleanLegacySubscriptionFields';
